requirements.txt
# File: requirements.txt
# Purpose: Python dependencies for anything-agents FastAPI microservice
# Dependencies: None
# Used by: pip install command for setting up the development environment

# =============================================================================
# WEB FRAMEWORK
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0

# =============================================================================
# HTTP CLIENT & ASYNC SUPPORT
# =============================================================================
httpx==0.25.2
aiofiles==23.2.1

# =============================================================================
# DATA VALIDATION & SERIALIZATION
# =============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# =============================================================================
# CACHING & STORAGE
# =============================================================================
redis==5.0.1

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing async HTTP calls

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
python-dotenv==1.0.0
openai-agents==0.1.0  # OpenAI Agents SDK for multiagent systems
tavily-python==0.3.3  # Tavily web search API client 